requirements.txt

pandas
scikit-learn
torch
kafka-python
psycopg2
redash-toolbelt
